+++ config/database.yml
@@ -1,0 +1,16 @@
+default_connection: default
+connections:
+  - default:
+      driver: mysql
+      version: 8.0
+      host: localhost
+      port: 3306
+      database: confucius
+      user: '%ENV:DB_USER%'
+      password: '%ENV:DB_PASSWORD%'
+      max_connections: 15
+      connection_lifetime: 30
+  - cache:
+      driver: "redis"
+      host: localhost
+      port: 6389
