+++ docker-compose.yaml
@@ -1,8 +1,8 @@
 version: '3'
 
 services:
-  backend:
-    container_name: "go_backend"
+  go:
+    container_name: backend
     tty: true
     image: golang:1.15-alpine
     volumes:
@@ -12,16 +12,17 @@ services:
     env_file:
       - ./.env
     depends_on:
-      - db
+      - mysql
       - redis
     ports:
       - 80:80
       - 443:443
 
-  db:
-    container_name: "mysql"
+  mysql:
+    container_name: database
     image: mysql:latest
     environment:
+      - MYSQL_DATABASE=confucius
       - MYSQL_ROOT_PASSWORD=secret
     command: --default-authentication-plugin=mysql_native_password
     restart: always
@@ -29,11 +30,11 @@ services:
       - 3306:3306
 
   redis:
-    container_name: "redis"
+    container_name: cache
     image: redis:alpine
 
-  monitoring:
-    container_name: "prometheus"
+  prometheus:
+    container_name: monitoring
     image: prom/prometheus:latest
     ports:
       - 9090:9090
