+++ service/db/service.go
@@ -1,0 +1,1 @@
+package db
