+++ service/server/config_test.go
@@ -1,0 +1,1 @@
+package server
