--- src/graphql/connection_queue.go
@@ -1,3 +1,0 @@
-package graphql
-
-// TODO: Database connection queue for asynchronous queries
