--- src/graphql/dbal.go
@@ -1,3 +1,0 @@
-package graphql
-
-// TODO: Implement abstraction level for MySQL
