--- src/graphql/drivers/mysql.go
@@ -1,3 +1,0 @@
-package drivers
-
-// TODO: MySQL driver for database abstraction layer
