--- src/graphql/migration.go
@@ -1,3 +1,0 @@
-package graphql
-
-// TODO: Implement GraphQL schema migration to database
