--- src/graphql/schema_resolver_test.go
@@ -1,1 +1,0 @@
-package graphql
