+++ utils/tasks.go
@@ -1,3 +1,3 @@
-package server
+package utils
 
 // TODO: Implement tasks similar to crontab
